Move image parsing to docker package
