+++ pkg/skaffold/deploy/kubectl.go
@@ -31,7 +31,6 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/docker"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/v1alpha2"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
-	"github.com/docker/distribution/reference"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
@@ -332,52 +331,22 @@ func recursiveReplaceImage(i interface{}, replacements map[string]*replacement) 
 			}
 
 			image := v.(string)
-			parsed, err := parseReference(image)
+			parsed, err := docker.ParseReference(image)
 			if err != nil {
 				logrus.Warnf("Couldn't parse image: %s", v)
 				continue
 			}
 
-			if parsed.fullyQualified {
+			if parsed.FullyQualified {
 				// TODO(1.0.0): Remove this warning.
 				logrus.Infof("Not replacing fully qualified image: %s (see #565)", v)
 				continue
 			}
 
-			if img, present := replacements[parsed.baseName]; present {
+			if img, present := replacements[parsed.BaseName]; present {
 				t[k] = img.tag
 				img.found = true
 			}
 		}
 	}
-}
-
-type imageReference struct {
-	baseName       string
-	fullyQualified bool
-}
-
-func parseReference(image string) (*imageReference, error) {
-	r, err := reference.Parse(image)
-	if err != nil {
-		return nil, err
-	}
-
-	baseName := image
-	if n, ok := r.(reference.Named); ok {
-		baseName = n.Name()
-	}
-
-	fullyQualified := false
-	switch n := r.(type) {
-	case reference.Tagged:
-		fullyQualified = n.Tag() != "latest"
-	case reference.Digested:
-		fullyQualified = true
-	}
-
-	return &imageReference{
-		baseName:       baseName,
-		fullyQualified: fullyQualified,
-	}, nil
 }
